requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-Login==0.6.2
Werkzeug==2.3.7
fpdf==1.7.2
qrcode==7.4.2
Pillow==10.1.0
python-dotenv==1.0.0
psycopg2-binary
gunicorn
